html/obsolete/requirements-for-implementations/other-elements-attributes-and-apis/document-all.html

<!DOCTYPE html>
<title>document.all</title>
<link rel="author" title="Corey Farwell" href="mailto:coreyf@rwell.org">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-all">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function test_all(all) {
  assert_false(Boolean(document.all));

  assert_true(document.all == undefined);
  assert_true(document.all == null);
  assert_false(document.all != undefined);
  assert_false(document.all != null);

  assert_true(document.all !== undefined);
  assert_true(document.all !== null);
  assert_false(document.all === undefined);
  assert_false(document.all === null);

  assert_equals(typeof document.all, "undefined");

  if (document.all) { assert_true(false); }

  if (!document.all) {}
  else { assert_true(false); }
}

test(function() {
  test_all(document.all);

  var all = document.all;
  test_all(all);
}, "'unusual behaviors'")
</script>
